==> notebooks/07_accessibility.ipynb <==
This script calculates the accessibilty for each green area

### 0. Libraries

import geopandas as gpd
import pandas as pd

### 1. Accessibility

# load the file
access_ga = gpd.read_file("data/Climate_Shelter_Index/isocrona/access_green/access_green.shp")

access_ga.head()

#### 1.1 Cycleway

# Identify cycleway
cycleway = access_ga[access_ga['highway']=='cycleway']
cycleway.head()

# Create a DataFrame to store the aggregation
sum_cycleway = cycleway[['nome', 'highway']].copy()
# Count the occurrences of each 'nome'
sum_cycleway['n_cycleway'] = sum_cycleway.groupby('nome')['nome'].transform('count')

sum_cycleway.head()

# Drop duplicates to keep only unique rows
sum_cycleway = sum_cycleway.drop_duplicates(subset=['nome', 'highway', 'n_cycleway'])
sum_cycleway.head()

sum_cycleway.to_csv('data\Climate_Shelter_Index\isocrona\sum_cycleway.csv', index=False)

#### 1.2 Footway

# Identify street accessible by foot
footway= access_ga[(access_ga['highway']=='footway')]
footway.head()

# Create a DataFrame to store the aggregation
sum_footway = footway[['nome', 'highway']].copy()
# Count the occurrences of each 'nome'
sum_footway['n_footway'] = sum_footway.groupby('nome')['nome'].transform('count')

sum_footway.head()

# Drop duplicates to keep only unique rows
sum_footway = sum_footway.drop_duplicates(subset=['nome', 'highway', 'n_footway'])
sum_footway.head()

sum_footway.to_csv('data\Climate_Shelter_Index\isocrona\sum_footway.csv', index=False)

#### 1.3 Footway + Cycleway

# Load  files
cycleway_csv = pd.read_csv('data\Climate_Shelter_Index\isocrona\sum_cycleway.csv')
footway_csv = pd.read_csv('data\Climate_Shelter_Index\isocrona\sum_footway.csv')
green_areas = gpd.read_file("data\Climate_Shelter_Index\green_areas__ndvi.shp")
green_areas.head(3)

# Merge the shapefile with the DataFrame about cycleway based on the 'nome' column
merged_df = green_areas.merge(cycleway_csv[['nome', 'n_cycleway']], on='nome', how='left')
# Convert 'n_footway' to integers and fill missing values with 0
merged_df['n_cycleway'] = merged_df['n_cycleway'].fillna(0).astype(int)

# Merge the shapefile with the DataFrame about footway based on the 'nome' column
merged_df = merged_df.merge(footway_csv[['nome', 'n_footway']], on='nome', how='left')
# Convert 'n_footway' to integers and fill missing values with 0
merged_df['n_footway'] = merged_df['n_footway'].fillna(0).astype(int)

merged_df.head()

# Save the shapefile
merged_df.to_file('data\Climate_Shelter_Index\isocrona\green_area_accessibility.shp')

#### 1.4 Footway + cycleway + bus stop

# Load  files
bus_5min = pd.read_csv('data/Climate_Shelter_Index/isocrona/sum_bus_5_stop.csv')
bus_10min = pd.read_csv('data/Climate_Shelter_Index/isocrona/sum_bus_10_stop.csv')
accessibility_ga = gpd.read_file('data/Climate_Shelter_Index/isocrona/green_area_accessibility.shp')

bus_10min.shape

# rename columns
bus_5min = bus_5min.rename(columns= {'n_busstop':'busstop5'})
bus_10min = bus_10min.rename(columns= {'n_busstop':'busstop10'})

## 5 min
# Merge the shapefile DataFrame with the CSV DataFrame based on the 'nome' column
merged_df = accessibility_ga.merge(bus_5min[['nome', 'busstop5']], on='nome', how='left')

# Convert 'busstop5' to integers and fill missing values with 0
merged_df['busstop5'] = merged_df['busstop5'].fillna(0).astype(int)

## 10 min
# Merge the shapefile DataFrame with the CSV DataFrame based on the 'nome' column
merged_df = merged_df.merge(bus_10min[['nome', 'busstop10']], on='nome', how='left')

# Convert 'busstop5' to integers and fill missing values with 0
merged_df['busstop10'] = merged_df['busstop10'].fillna(0).astype(int)

merged_df.head()

# Save the shapefile
merged_df.to_file('data\Climate_Shelter_Index\isocrona\green_area_accessibility_bus.shp')

#### 1.5 Footway + cycleway + bus stop + disabled parking

# Load  files
disab_parking = pd.read_csv('data/Climate_Shelter_Index/isocrona/disabled_parking.csv')
accessibility_bus_ga = gpd.read_file('data/Climate_Shelter_Index/isocrona/green_area_accessibility_bus.shp')

disab_parking

# Merge the shapefile DataFrame with the GeoDataFrame based on the 'nome' column
merged_df = accessibility_bus_ga.merge(disab_parking, on='nome', how='left')

# Convert 'disabled_p' to integers and fill missing values with 0
merged_df['disab_p_5'] = merged_df['disab_p_5'].fillna(0).astype(int)

# Replace 0 with -1 in the 'disab_p_5' column to assign a negative value if there no disabled parking
merged_df['disab_p_5'] = merged_df['disab_p_5'].replace(0, -1)

merged_df.head()

# Save the shapefile
merged_df.to_file('data\Climate_Shelter_Index\isocrona\complete_accessibility')

### 2. Index of Accessibility
Consider some aspects to calculate the index: 
- the park is accessible via a foot path (how many accesses?)
- the park is accessible via a bike path (how many accesses?)
- the park has a bus stop nearby (hoe many bus stop within a 5-minute walk? or 10 minute walk?)


# Load file
complete_accessibility = gpd.read_file('data/Climate_Shelter_Index/isocrona/complete_accessibility/complete_accessibility.shp')

# Define the weights for each aspect
weight_cycleway = 0.3
weight_footway = 0.3
weight_busstop5 = 0.2
weight_busstop10 = 0.1
weight_disabled_parking = 0.1

# Normalize and calculate the accessibility index for each aspect
def normalize_and_calculate_index(col_name, weight):
    min_value = complete_accessibility[col_name].min()
    max_value = complete_accessibility[col_name].max()
    normalized_value = (complete_accessibility[col_name] - min_value) / (max_value - min_value)
    return normalized_value * weight

# Calculate the accessibility index for each aspect
complete_accessibility['index_n_cycleway'] = normalize_and_calculate_index('n_cycleway', weight_cycleway)
complete_accessibility['index_n_footway'] = normalize_and_calculate_index('n_footway', weight_footway)
complete_accessibility['index_busstop5'] = normalize_and_calculate_index('busstop5', weight_busstop5)
complete_accessibility['index_busstop10'] = normalize_and_calculate_index('busstop10', weight_busstop10)
complete_accessibility['index_disab_p_5'] = normalize_and_calculate_index('disab_p_5', weight_disabled_parking)

# Calculate the total_accessibility_index
complete_accessibility['access_ind'] = complete_accessibility[['index_n_cycleway', 'index_n_footway', 'index_busstop5', 'index_busstop10', 'index_disab_p_5']].sum(axis=1)


# Remove the individual index columns
complete_accessibility.drop(columns=['index_n_cycleway', 'index_n_footway', 'index_busstop5', 'index_busstop10', 'index_disab_p_5'], inplace=True)

complete_accessibility.head()

complete_accessibility.shape

# Save the shapefile
complete_accessibility.to_file('data\Climate_Shelter_Index\isocrona\access_index')
