==> notebooks/09_new_CSI.ipynb <==
This script calculates the Climate Shelter Index for each public green area.  
A climate shelter should have these characteristics:
- Surface area > 0.5 ha 
- NDVI > 0.4 
- Good accessibility
- Presence of drinking fountains 
- Presence of  benches 
- Presence of picnic tables

### 0. Libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
import geopandas as gpd

### 1. Climate Shelter Index - CSI

# Load your shapefile
green_area = gpd.read_file('data\Climate_Shelter_Index\isocrona\access_index')

green_area.head()

green_area.columns

#### 1.1 Determine weights with PCA

# Assuming 'id_df', 'id_bench', and 'id_pt' contain lists of ids, we will count them.

green_area['n_df'] = green_area['id_df'].apply(lambda x: len(x.split(',')) if pd.notna(x) else 0)
green_area['n_bench'] = green_area['id_bench'].apply(lambda x: len(x.split(',')) if pd.notna(x) else 0)
green_area['n_pt'] = green_area['id_pt'].apply(lambda x: len(x.split(',')) if pd.notna(x) else 0)
green_area['norm_area'] = green_area['area_Ha'].apply(lambda x: x if pd.notna(x) else 0) # substitue NA with 0
green_area['norm_area'] = green_area['norm_area'].apply(lambda x: x if x > 0 else 0)
green_area['norm_NDVI'] = green_area['NDVI'].apply(lambda x: x if pd.notna(x) else 0)   # substitue NA with 0

# Print to verify the new columns
print(green_area[[ 'n_df', 'n_bench', 'n_pt', 'norm_area', 'norm_NDVI']].head() )      

# Step 1: Ensure that only necessary columns are used
pca_green_area = green_area[['n_df', 'n_bench', 'n_pt', 'norm_area', 'norm_NDVI', 'access_ind']]

# Step 2: Standardize the data
scaler = StandardScaler()
scaled_data = scaler.fit_transform(pca_green_area[['n_df', 'n_bench', 'n_pt', 'norm_area', 'norm_NDVI', 'access_ind']])

# Step 3: Perform PCA
pca = PCA()
pca.fit(scaled_data)

# Step 4: Analyze PCA output
print("Variance explained by each component:", pca.explained_variance_ratio_)
print("Cumulative variance explained:", np.cumsum(pca.explained_variance_ratio_))

# Step 5: Extract and visualize PCA loadings
loadings = pca.components_.T * np.sqrt(pca.explained_variance_)


# Step 6: create df with loadings
loadings_df = pd.DataFrame(loadings, columns=[f'PC{i+1}' for i in range(len(pca_green_area.columns))], index=pca_green_area.columns)
print(loadings_df)

#Step 7: Calculate normalized weights
weights = np.abs(loadings_df['PC1']) / np.sum(np.abs(loadings_df['PC1']))
print(weights)

# Define weights for each characteristic (surface area, NDVI, accessibility, drinking fountains, benches, picnic tables)
weights_dict = weights.round(2).to_dict()
print(weights_dict)

# Define the columns you want to normalize
columns_to_normalize = ['n_df', 'n_bench', 'n_pt']

# Perform min-max normalization for each column
for col_name in columns_to_normalize:
    min_value = green_area[col_name].min()
    max_value = green_area[col_name].max()
    green_area[col_name] = (green_area[col_name] - min_value) / (max_value - min_value)


# Normalize surface area (0.5 ha threshold)
green_area['norm_area'] = green_area['norm_area'].apply(lambda x: 1 if x >= 0.5 else 0)


# Normalize NDVI (0.4 threshold)
green_area['norm_NDVI'] = green_area['norm_NDVI'].apply(lambda x: x if x >= 0.4 else 0)


# Calculate the CSI based on the weighted and normalized variables
green_area['CSI'] = (
    weights_dict['norm_area'] * green_area['norm_area'] +
    weights_dict['norm_NDVI'] * green_area['norm_NDVI'] +
    weights_dict['access_ind'] * green_area['access_ind'] +
    weights_dict['n_df'] * green_area['n_df'] +
    weights_dict['n_bench'] * green_area['n_bench'] +
    weights_dict['n_pt'] * green_area['n_pt']
)

green_area = green_area.drop(columns=['n_df', 'n_bench', 'n_pt', 'norm_area', 'norm_NDVI'])

green_area[green_area['NDVI']<0.41].head()

green_area.to_file('data\Climate_Shelter_Index\isocrona\CSI')

#### 1.2 CSI for selected areas


csi = gpd.read_file('data\Climate_Shelter_Index\isocrona\CSI')
covered_areas = gpd.read_file("data\Meteoblue\covered_areas.shp")

# select only green area within covered areas
covered_areas_csi =csi[csi.geometry.within(covered_areas.unary_union)] 

covered_areas_csi.columns

covered_areas_csi_sorted = covered_areas_csi.sort_values(by='CSI', ascending=False).head(5)

for index, row in covered_areas_csi_sorted.iterrows():
    print(f"{row['nome']}, with CSI of {round(row['CSI'],3)}")

covered_areas_csi_sorted.explore()

covered_areas_csi.to_file('data\Climate_Shelter_Index\CSI_covered_areas')

covered_areas_csi_sorted.to_file('data\Climate_Shelter_Index\CSI_covered_areas_top5')

##### 1.2.1 Green Surface > 0.5

covered_areas_csi[covered_areas_csi['area_Ha']>0.5].shape

# find the top 10 parks with the largest surface area
covered_areas_csi_sorted_2 = covered_areas_csi.sort_values(by='area_Ha', ascending=False).head(10)

for index, row in covered_areas_csi_sorted_2.iterrows():
    print(f"{row['nome']}, with a green area of {round(row['area_Ha'], 3)} hectares")

##### 1.2.2 NDVI > 0.4

covered_areas_csi[covered_areas_csi['NDVI']>0.4].shape

##### 1.2.3 NDVI + Green Surface

covered_areas_csi[(covered_areas_csi['NDVI']>0.4) & (covered_areas_csi['area_Ha']>0.5)].shape
